# control-plane/internal/api/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Ascend 虚拟化资源池控制面 API
  version: "0.1.0"
  description: |
    与前端字段契约（v0）：JSON 字段名均为 snake_case；时间 RFC3339 UTC。
    鉴权：`Authorization: Bearer <token>`（`CP_AUTH_DISABLED=1` 时跳过，等价 admin）。
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/workloads:
    get:
      summary: 任务列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: tenant
          schema: { type: string }
        - in: query
          name: phase
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  workloads:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkloadRecord"
...
  /api/v1/workloads/{id}:
    get:
      summary: 任务详情
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkloadRecord"
    delete:
      summary: 删除任务与集群对象
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: CP1-HMAC
  schemas:
...
    WorkloadRecord:
      type: object
      properties:
        id: { type: string }
        tenant: { type: string }
        namespace: { type: string }
        name: { type: string }
        kind: { type: string }
        phase: { type: string }
        flavor_id: { type: string }
        extended_resource: { type: string }
        message: { type: string }
        events_summary:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...